Materialize into a custom ListMultimap
We know we won't have duplicate instances, so there's no reason to use a
set, and this should be cheaper and more compact.  We prune eagerly and
avoid repeated pruning if we hit in the materialization-sharing cache.
Previously we copied the result of Multimaps.index purely to let us
remove when pruning, which is now unnecessary.  The streams machinery
for counting devices wasn't really any terser and costs performance.